Expose VLLM/DeepSpeed flags and guard LLM imports
Move LLM-only imports behind TYPE_CHECKING/HAS_LLM_DEPENDENCIES to avoid
hard runtime dependencies. Add HAS_VLLM and HAS_DEEPSPEED flags to
package
init, add __future__ annotations, and tuck PEFT/vllm/deepspeed imports
so
typing still works. Introduce a thin Liger DPO wrapper exposing alpha,
adjust PeftModel typing fallback, and update tests for the changes.
