+++ agilerl/__init__.py
@@ -27,5 +27,8 @@ def _is_distribution_installed(distribution: str) -> bool:
     return True
 
 
+# Use these flags for laxzy import checks
 HAS_LLM_DEPENDENCIES = all(_is_distribution_installed(pkg) for pkg in LLM_PACKAGES)
 HAS_LIGER_KERNEL = _is_distribution_installed("liger-kernel")
+HAS_VLLM = _is_distribution_installed("vllm")
+HAS_DEEPSPEED = _is_distribution_installed("deepspeed")
