+++ agilerl/algorithms/dpo.py
@@ -1,19 +1,91 @@
+from __future__ import annotations
+
 import gc
 import warnings
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 import numpy as np
 import torch
 import torch.nn.functional as F
-from accelerate import Accelerator
 
 from agilerl import HAS_LIGER_KERNEL
+
+if TYPE_CHECKING:
+    from accelerate import Accelerator
+    from peft import LoraConfig
+
 from agilerl.algorithms.core.base import LLMAlgorithm
 from agilerl.algorithms.core.registry import HyperparameterConfig, NetworkGroup
-from agilerl.protocols import LoraConfigProtocol, PreTrainedModelProtocol
+from agilerl.protocols import PreTrainedModelProtocol
 from agilerl.typing import ExperiencesType, LLMObsType
 from agilerl.utils.algo_utils import get_experiences_samples
-from agilerl.utils.llm_utils import PreferenceGym, _LigerDPOWithAlpha
+from agilerl.utils.llm_utils import PreferenceGym
+
+if HAS_LIGER_KERNEL or TYPE_CHECKING:
+    from liger_kernel.chunked_loss.dpo_loss import LigerFusedLinearDPOFunction
+    from liger_kernel.chunked_loss.fused_linear_preference import (
+        LigerFusedLinearPreferenceBase,
+    )
+
+    class _LigerDPOWithAlpha(LigerFusedLinearPreferenceBase):
+        """Thin wrapper that exposes ``alpha`` for NLL scaling.
+
+        ``LigerFusedLinearDPOFunction`` passes ``compute_nll_loss`` as a bool
+        but never forwards ``alpha`` to the base class (which defaults to 1.0).
+        This subclass reuses the DPO preference loss and adds ``alpha`` so the
+        fused kernel correctly scales the NLL component.
+        """
+
+        preference_loss_fn = staticmethod(
+            LigerFusedLinearDPOFunction.preference_loss_fn
+        )
+
+        @classmethod
+        def forward(
+            cls,
+            ctx,
+            _input,
+            weight,
+            target,
+            bias=None,
+            ref_input=None,
+            ref_weight=None,
+            ref_bias=None,
+            ignore_index=-100,
+            beta=0.1,
+            alpha=1.0,
+            compute_nll_loss=True,
+            compiled=True,
+            use_ref_model=True,
+            average_log_prob=False,
+            chunk_size=1,
+            loss_type="sigmoid",
+        ):
+            return LigerFusedLinearPreferenceBase.forward(
+                cls=cls,
+                ctx=ctx,
+                _input=_input,
+                weight=weight,
+                target=target,
+                bias=bias,
+                ignore_index=ignore_index,
+                alpha=alpha,
+                beta=beta,
+                compute_nll_loss=compute_nll_loss,
+                compiled=compiled,
+                use_ref_model=use_ref_model,
+                ref_input=ref_input,
+                ref_weight=ref_weight,
+                ref_bias=ref_bias,
+                average_log_prob=average_log_prob,
+                chunk_size=chunk_size,
+                loss_type=loss_type,
+            )
+
+        @staticmethod
+        def backward(ctx, *grad_output):
+            grads = LigerFusedLinearPreferenceBase.backward(ctx, grad_output)[:4]
+            return (*grads, *(None,) * 12)
 
 
 class DPO(LLMAlgorithm):
@@ -54,7 +126,7 @@ class DPO(LLMAlgorithm):
     :param device: Device for accelerated computing, 'cpu' or 'cuda', defaults to 'cpu'
     :type device: str, optional
     :param lora_config: Config for LoRA, defaults to None
-    :type lora_config: LoraConfigProtocol, optional
+    :type lora_config: LoraConfig, optional
     :param accelerator: Accelerator for distributed computing, defaults to None
     :type accelerator: accelerate.Accelerator(), optional
     :param wrap: Wrap models for distributed training upon creation, defaults to True
@@ -92,7 +164,7 @@ class DPO(LLMAlgorithm):
         micro_batch_size_per_gpu: int | None = None,
         reduce_memory_peak: bool = False,
         device: str = "cpu",
-        lora_config: LoraConfigProtocol | None = None,
+        lora_config: LoraConfig | None = None,
         accelerator: Accelerator | None = None,
         wrap: bool = True,
         clone: bool = False,
