+++ agilerl/utils/llm_utils.py
@@ -11,95 +11,19 @@ from typing import Any
 import gymnasium as gym
 import torch
 from accelerate import Accelerator
+from datasets import Dataset
 from torch import nn
 from torch.utils.data import DataLoader
+from transformers import AutoModelForCausalLM, AutoTokenizer
+from transformers.modeling_utils import PreTrainedModel
+from transformers.tokenization_utils_base import BatchEncoding
 
-from agilerl import HAS_LIGER_KERNEL, HAS_LLM_DEPENDENCIES
 from agilerl.typing import PreferencePrompts, ReasoningPrompts, SFTPrompts
 
-if HAS_LLM_DEPENDENCIES:
-    try:
-        import deepspeed
-    except ImportError:
-        deepspeed = None
-    if HAS_LIGER_KERNEL:
-        from liger_kernel.chunked_loss.dpo_loss import LigerFusedLinearDPOFunction
-        from liger_kernel.chunked_loss.fused_linear_preference import (
-            LigerFusedLinearPreferenceBase,
-        )
-
-        class _LigerDPOWithAlpha(LigerFusedLinearPreferenceBase):
-            """Thin wrapper that exposes ``alpha`` for NLL scaling.
-
-            ``LigerFusedLinearDPOFunction`` passes ``compute_nll_loss`` as a bool
-            but never forwards ``alpha`` to the base class (which defaults to 1.0).
-            This subclass reuses the DPO preference loss and adds ``alpha`` so the
-            fused kernel correctly scales the NLL component.
-            """
-
-            preference_loss_fn = staticmethod(
-                LigerFusedLinearDPOFunction.preference_loss_fn
-            )
-
-            @classmethod
-            def forward(
-                cls,
-                ctx,
-                _input,
-                weight,
-                target,
-                bias=None,
-                ref_input=None,
-                ref_weight=None,
-                ref_bias=None,
-                ignore_index=-100,
-                beta=0.1,
-                alpha=1.0,
-                compute_nll_loss=True,
-                compiled=True,
-                use_ref_model=True,
-                average_log_prob=False,
-                chunk_size=1,
-                loss_type="sigmoid",
-            ):
-                return LigerFusedLinearPreferenceBase.forward(
-                    cls=cls,
-                    ctx=ctx,
-                    _input=_input,
-                    weight=weight,
-                    target=target,
-                    bias=bias,
-                    ignore_index=ignore_index,
-                    alpha=alpha,
-                    beta=beta,
-                    compute_nll_loss=compute_nll_loss,
-                    compiled=compiled,
-                    use_ref_model=use_ref_model,
-                    ref_input=ref_input,
-                    ref_weight=ref_weight,
-                    ref_bias=ref_bias,
-                    average_log_prob=average_log_prob,
-                    chunk_size=chunk_size,
-                    loss_type=loss_type,
-                )
-
-            @staticmethod
-            def backward(ctx, *grad_output):
-                grads = LigerFusedLinearPreferenceBase.backward(ctx, grad_output)[:4]
-                return (*grads, *(None,) * 12)
-    else:
-        _LigerDPOWithAlpha = None
-    from datasets import Dataset
-    from transformers import AutoModelForCausalLM, AutoTokenizer
-    from transformers.modeling_utils import PreTrainedModel
-    from transformers.tokenization_utils_base import BatchEncoding
-else:
+try:
+    import deepspeed
+except ImportError:
     deepspeed = None
-    _LigerDPOWithAlpha = None
-    AutoTokenizer = Any
-    PreTrainedModel = Any
-    BatchEncoding = Any
-    Dataset = Any
 
 
 def apply_chat_template(
