+++ tests/test_algorithms/test_llms/test_sft.py
@@ -484,7 +484,6 @@ def test_sft_liger_unavailable_behaviour(
     assertion_mode,
 ):
     monkeypatch.setattr("agilerl.algorithms.sft.HAS_LIGER_KERNEL", False)
-    monkeypatch.setattr("agilerl.algorithms.sft.LigerCrossEntropyLoss", None)
     if assertion_mode == "warns_and_fallback":
         with pytest.warns(
             UserWarning,
